Move ReplClient
